+++ CHANGELOG.md
@@ -59,3 +59,4 @@
 - [2026-04-02 13:03] feat(cache): add bi-directional event bus handler for cache
 - [2026-04-02 13:13] docs(theme): update readme and API contracts for theme
 - [2026-04-02 13:16] test(analytics): implement widget integration tests for analytics
+- [2026-04-02 13:24] feat(theme): integrate custom telemetry logger in theme